refact: autopep8 & .gitignore & datasets
1. datasets 폴더 지우지마셈 (원래 .gitignore 파일 있어서 데이터셋은 추가 안됨)
2. 2 spaces --> 4 spaces
3. .gitignore update (osx, pycharm, ...)
4. spring1617_assignment1/assignment1/... 중첩폴더 삭제
