+++ assignment1/collectSubmission.sh
