+++ assignment1/cs231n/classifiers/k_nearest_neighbor.py
@@ -68,7 +68,8 @@ class KNearestNeighbor(object):
         dists = np.zeros((num_test, num_train))
         for i in xrange(num_test):
             for j in xrange(num_train):
-                dists[i, j] = np.sqrt(np.sum((X[i, :] - self.X_train[j, :])**2))
+                dists[i, j] = np.sqrt(
+                    np.sum((X[i, :] - self.X_train[j, :])**2))
         return dists
 
     def compute_distances_one_loop(self, X):
@@ -84,14 +85,14 @@ class KNearestNeighbor(object):
         for i in xrange(num_test):
             dists[i, :] = np.sqrt(np.sum((X[i, :] - self.X_train)**2, axis=1))
 
-            #######################################################################
+            ###################################################################
             # TODO:                                                               #
             # Compute the l2 distance between the ith test point and all training #
             # points, and store the result in dists[i, :].                        #
-            #######################################################################
-            #######################################################################
+            ###################################################################
+            ###################################################################
             #                         END OF YOUR CODE                            #
-            #######################################################################
+            ###################################################################
         return dists
 
     def compute_distances_no_loops(self, X):
@@ -104,7 +105,7 @@ class KNearestNeighbor(object):
         num_test = X.shape[0]
         num_train = self.X_train.shape[0]
         dists = np.zeros((num_test, num_train))
-        #########################################################################
+        #######################################################################
         # TODO:                                                                 #
         # Compute the l2 distance between all test points and all training      #
         # points without using any explicit loops, and store the result in      #
@@ -115,15 +116,15 @@ class KNearestNeighbor(object):
         #                                                                       #
         # HINT: Try to formulate the l2 distance using matrix multiplication    #
         #       and two broadcast sums.                                         #
-        #########################################################################
-        dot = -2*np.dot(X, np.transpose(self.X_train))
+        #######################################################################
+        dot = -2 * np.dot(X, np.transpose(self.X_train))
         train = np.sum(self.X_train**2, axis=1)
         test = np.sum(X**2, axis=1)
         temp = dot + train
         dists = np.sqrt(np.transpose(np.transpose(temp) + test))
-        #########################################################################
+        #######################################################################
         #                         END OF YOUR CODE                              #
-        #########################################################################
+        #######################################################################
         return dists
 
     def predict_labels(self, dists, k=1):
@@ -145,26 +146,25 @@ class KNearestNeighbor(object):
             # A list of length k storing the labels of the k nearest neighbors to
             # the ith test point.
             closest_y = []
-            #########################################################################
+            ###################################################################
             # TODO:                                                                 #
             # Use the distance matrix to find the k nearest neighbors of the ith    #
             # testing point, and use self.y_train to find the labels of these       #
             # neighbors. Store these labels in closest_y.                           #
             # Hint: Look up the function numpy.argsort.                             #
-            #########################################################################
-            closest_y = self.y_train[np.argsort(dists[i,:])[:k]]
-            #########################################################################
+            ###################################################################
+            closest_y = self.y_train[np.argsort(dists[i, :])[:k]]
+            ###################################################################
             # TODO:                                                                 #
             # Now that you have found the labels of the k nearest neighbors, you    #
             # need to find the most common label in the list closest_y of labels.   #
             # Store this label in y_pred[i]. Break ties by choosing the smaller     #
             # label.                                                                #
-            #########################################################################
+            ###################################################################
             y_pred[i] = np.argmax(np.bincount(closest_y))
 
-
-            #########################################################################
+            ###################################################################
             #                           END OF YOUR CODE                            #
-            #########################################################################
+            ###################################################################
 
         return y_pred
