+++ assignment1/cs231n/classifiers/linear_classifier.py
@@ -43,7 +43,7 @@ class LinearClassifier(object):
             X_batch = None
             y_batch = None
 
-            #########################################################################
+            ###################################################################
             # TODO:                                                                 #
             # Sample batch_size elements from the training data and their           #
             # corresponding labels to use in this round of gradient descent.        #
@@ -53,28 +53,28 @@ class LinearClassifier(object):
             #                                                                       #
             # Hint: Use np.random.choice to generate indices. Sampling with         #
             # replacement is faster than sampling without replacement.              #
-            #########################################################################
+            ###################################################################
             sample_array = np.random.choice(num_train, batch_size)
-            
+
             X_batch = X[sample_array, :]
             y_batch = y[sample_array]
-            #########################################################################
+            ###################################################################
             #                       END OF YOUR CODE                                #
-            #########################################################################
+            ###################################################################
 
             # evaluate loss and gradient
             loss, grad = self.loss(X_batch, y_batch, reg)
             loss_history.append(loss)
 
             # perform parameter update
-            #########################################################################
+            ###################################################################
             # TODO:                                                                 #
             # Update the weights using the gradient and the learning rate.          #
-            #########################################################################
+            ###################################################################
             self.W = self.W - learning_rate * grad
-            #########################################################################
+            ###################################################################
             #                       END OF YOUR CODE                                #
-            #########################################################################
+            ###################################################################
 
             if verbose and it % 100 == 0:
                 print('iteration %d / %d: loss %f' % (it, num_iters, loss))
@@ -96,14 +96,14 @@ class LinearClassifier(object):
           class.
         """
         y_pred = np.zeros(X.shape[0])
-        ###########################################################################
+        #######################################################################
         # TODO:                                                                   #
         # Implement this method. Store the predicted labels in y_pred.            #
-        ###########################################################################
+        #######################################################################
         y_pred = np.argmax(X.dot(self.W), axis=1)
-        ###########################################################################
+        #######################################################################
         #                           END OF YOUR CODE                              #
-        ###########################################################################
+        #######################################################################
         return y_pred
 
     def loss(self, X_batch, y_batch, reg):
