+++ assignment1/cs231n/datasets/.gitignore
@@ -1,0 +1,4 @@
+cifar-10-batches-py/*
+tiny-imagenet-100-A*
+tiny-imagenet-100-B*
+tiny-100-A-pretrained/*
