+++ assignment1/cs231n/datasets/get_datasets.sh
@@ -1,0 +1,4 @@
+# Get CIFAR10
+wget http://www.cs.toronto.edu/~kriz/cifar-10-python.tar.gz
+tar -xzvf cifar-10-python.tar.gz
+rm cifar-10-python.tar.gz
