+++ assignment1/setup_googlecloud.sh
