+++ assignment1/start_ipython_osx.sh
